+++ app/code/community/Nosto/Tagging/Block/Adminhtml/System/Config/Inventory/Switch.php
@@ -1,9 +1,9 @@
 <?php
 /**
  * Magento
- *  
+ *
  * NOTICE OF LICENSE
- *  
+ *
  * This source file is subject to the Open Software License (OSL 3.0)
  * that is bundled with this package in the file LICENSE.txt.
  * It is also available through the world-wide-web at this URL:
@@ -11,13 +11,13 @@
  * If you did not receive a copy of the license and are unable to
  * obtain it through the world-wide-web, please send an email
  * to license@magentocommerce.com so we can send you a copy immediately.
- *  
+ *
  * DISCLAIMER
- *  
+ *
  * Do not edit or add to this file if you wish to upgrade Magento to newer
  * versions in the future. If you wish to customize Magento for your
  * needs please refer to http://www.magentocommerce.com for more information.
- *  
+ *
  * @category  Nosto
  * @package   Nosto_Tagging
  * @author    Nosto Solutions Ltd <magento@nosto.com>
@@ -26,43 +26,34 @@
  */
 
 /**
- * Search term tagging block.
- * Adds search term tag to the HTML document.
+ * Extension system setting source model for enabling/disabling inventory level sync
  *
  * @category Nosto
  * @package  Nosto_Tagging
  * @author   Nosto Solutions Ltd <magento@nosto.com>
  */
-class Nosto_Tagging_Block_Searchterm extends Mage_Core_Block_Template
+class Nosto_Tagging_Block_Adminhtml_System_Config_Inventory_Switch extends Mage_Adminhtml_Block_System_Config_Form_Field
 {
     /**
-     * Render shopping search term content
-     *
+     * @param Varien_Data_Form_Element_Abstract $element
      * @return string
      */
-    protected function _toHtml()
+    protected function _getElementHtml(Varien_Data_Form_Element_Abstract $element)
     {
-        /** @var Nosto_Tagging_Helper_Account $helper */
-        $helper = Mage::helper('nosto_tagging/account');
-        if (!$helper->existsAndIsConnected()
-            || $this->getSearchTerm() === null
-            || !Mage::helper('nosto_tagging/module')->isModuleEnabled()
-        ) {
-            return '';
+        /** @var Nosto_Tagging_Helper_Data $helper */
+        $helper = Mage::helper('nosto_tagging');
+        $store = $helper->getStore();
+        /** @var Nosto_Tagging_Helper_Data $dataHelper */
+        $dataHelper = Mage::helper('nosto_tagging');
+        if (!$dataHelper->getUseInventoryLevel($store)) {
+            /** @noinspection PhpUndefinedMethodInspection */
+            $element->setDisabled('disabled');
+            $element->setData('value', 0);
+            $comment = 'Sending inventory level after purchase feature is inactive because inventory sync is disabled.';
+            $element->setData(
+                'comment', $comment
+            );
         }
-        $searchTerm = new Nosto_Object_SearchTerm($this->escapeHtml($this->getSearchTerm()));
-        return $searchTerm->toHtml();
-    }
-
-    /**
-     * Returns the search term
-     *
-     * @return string
-     */
-    public function getSearchTerm()
-    {
-        /** @var Mage_CatalogSearch_Helper_Data $helper */
-        $helper = $this->helper('catalogsearch');
-        return $helper->getQueryText();
+        return parent::_getElementHtml($element);
     }
 }