+++ app/code/community/Nosto/Tagging/Helper/Data.php
@@ -72,6 +72,11 @@ class Nosto_Tagging_Helper_Data extends Mage_Core_Helper_Abstract
     const XML_PATH_VARIATION_SWITCH = 'nosto_tagging/variation/switch';
 
     /**
+     * Path to store config of sending inventory level
+     */
+    const XML_PATH_SEND_INVENTORY_LEVEL_AFTER_PURCHASE = 'nosto_tagging/general/send_inventory_level_after_purchase';
+
+    /**
      * Path to store config scheduled currency exchange rate update enabled setting.
      */
     const XML_PATH_SCHEDULED_CURRENCY_EXCHANGE_RATE_UPDATE_ENABLED
@@ -141,6 +146,11 @@ class Nosto_Tagging_Helper_Data extends Mage_Core_Helper_Abstract
      * Path to store config for tagging low stock
      */
     const XML_PATH_USE_LOW_STOCK = 'nosto_tagging/general/use_low_stock';
+
+    /**
+     * Path to setting if personalized category sorting is enabled
+     */
+    const XML_PATH_USE_PERSONALIZED_CATEGORY_SORTING = 'nosto_tagging/general/personalized_category_sorting';
 
     /**
      * @var boolean the path for setting for product urls
@@ -307,7 +317,7 @@ class Nosto_Tagging_Helper_Data extends Mage_Core_Helper_Abstract
      *
      * @param Mage_Core_Model_Store|null $store the store model or null.
      *
-     * @return bool
+     * @return boolean
      */
     public function getUsePrettyProductUrls($store = null)
     {
@@ -503,7 +513,7 @@ class Nosto_Tagging_Helper_Data extends Mage_Core_Helper_Abstract
      * Returns on/off setting for product API
      *
      * @param Mage_Core_Model_Store|null $store the store model or null.
-     * @return bool
+     * @return boolean
      */
     public function getUseProductApi($store = null)
     {
@@ -547,7 +557,7 @@ class Nosto_Tagging_Helper_Data extends Mage_Core_Helper_Abstract
      * Returns on/off setting for SKUs
      *
      * @param Mage_Core_Model_Store|null $store the store model or null.
-     * @return bool
+     * @return boolean
      */
     public function getUseSkus($store = null)
     {
@@ -569,7 +579,7 @@ class Nosto_Tagging_Helper_Data extends Mage_Core_Helper_Abstract
      * Returns on/off setting for alternate image urls
      *
      * @param Mage_Core_Model_Store|null $store the store model or null.
-     * @return bool
+     * @return boolean
      */
     public function getUseAlternateImages($store = null)
     {
@@ -580,7 +590,7 @@ class Nosto_Tagging_Helper_Data extends Mage_Core_Helper_Abstract
      * Returns on/off setting for inventory level
      *
      * @param Mage_Core_Model_Store|null $store the store model or null.
-     * @return bool
+     * @return boolean
      */
     public function getUseInventoryLevel($store = null)
     {
@@ -588,10 +598,21 @@ class Nosto_Tagging_Helper_Data extends Mage_Core_Helper_Abstract
     }
 
     /**
+     * Returns on/off setting for sending inventory level after purchase
+     *
+     * @param Mage_Core_Model_Store|null $store the store model or null.
+     * @return boolean
+     */
+    public function getSendInventoryLevelAfterPurchase($store = null)
+    {
+        return (bool)Mage::getStoreConfig(self::XML_PATH_SEND_INVENTORY_LEVEL_AFTER_PURCHASE, $store);
+    }
+
+    /**
      * Returns on/off setting for using low stock
      *
      * @param Mage_Core_Model_Store|null $store the store model or null.
-     * @return bool
+     * @return boolean
      */
     public function getUseLowStock($store = null)
     {
@@ -607,6 +628,17 @@ class Nosto_Tagging_Helper_Data extends Mage_Core_Helper_Abstract
     public function getSendAddToCartEvent($store)
     {
         return (bool)Mage::getStoreConfig(self::XML_PATH_SEND_ADD_TO_CART_EVENT, $store);
+    }
+
+    /**
+     * Returns on/off setting for using personalized category sorting
+     *
+     * @param Mage_Core_Model_Store|null $store the store model or null.
+     * @return boolean
+     */
+    public function getUsePersonalizedCategorySorting($store = null)
+    {
+        return (bool)Mage::getStoreConfig(self::XML_PATH_USE_PERSONALIZED_CATEGORY_SORTING, $store);
     }
 
     /**