+++ app/code/community/Nosto/Tagging/Model/Meta/Product.php
@@ -116,7 +116,7 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
         }
         $brandAttribute = $dataHelper->getBrandAttribute($store);
         if ($product->hasData($brandAttribute)) {
-            $this->setBrand($this->getAttributeValue($product, $brandAttribute));
+            $this->setBrand($this->getAttributeValue($product, $brandAttribute, $store->getId()));
         }
         if (($tags = $this->buildTags($product, $store)) !== array()) {
             $this->setTag1($tags);
@@ -133,14 +133,13 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
         if ($dataHelper->getUseSkus($store)) {
             $this->amendSkus($product, $store);
         }
-
         if ($dataHelper->isMultiCurrencyMethodExchangeRate($store)) {
             $this->setVariationId($store->getBaseCurrencyCode());
         } elseif ($dataHelper->isVariationEnabled($store)) {
             $this->amendVariations($product, $store);
         }
         if ($dataHelper->getUseCustomFields($store)) {
-            $this->setCustomFields($this->loadCustomFields($product));
+            $this->setCustomFields($this->buildCustomFields($product, $store));
         }
 
         Mage::dispatchEvent(
@@ -366,7 +365,7 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
                     continue;
                 }
                 try {
-                    $attributeValue = $this->getAttributeValue($product, $key);
+                    $attributeValue = $this->getAttributeValue($product, $key, $store->getId());
                     if (empty($attributeValue)) {
                         continue;
                     }
@@ -506,22 +505,40 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
      *
      * @param Mage_Catalog_Model_Product $product
      * @param string $attributeName
+     * @param null|int $storeId The id of the current store
      * @return string|null
      * @suppress PhanUndeclaredMethod
+     * @suppress PhanTypeMismatchArgument
      */
-    protected function getAttributeValue(Mage_Catalog_Model_Product $product, $attributeName)
+    protected function getAttributeValue(
+        Mage_Catalog_Model_Product $product,
+        $attributeName,
+        $storeId = null
+    )
     {
         $attribute = $product->getResource()->getAttribute($attributeName);
         if ($attribute instanceof Mage_Catalog_Model_Resource_Eav_Attribute) {
             /** @noinspection PhpParamsInspection */
-            $attributeValue = $product->getAttributeText($attributeName);
+            if ($storeId && method_exists($product, 'setStoreId')) {
+                $product->setStoreId($storeId);
+            }
+            try {
+                /** @noinspection PhpUndefinedMethodInspection */
+                $attributeValue = $product->getResource()
+                    ->getAttribute($attribute)
+                    ->setStoreId($storeId)
+                    ->getSource()
+                    ->getOptionText($product->getData($attributeName));
+            } catch (\Exception $e) {
+                Nosto_Tagging_Helper_Log::exception($e);
+            }
             if (empty($attributeValue)) {
                 $attributeValue = $product->getData($attributeName);
             }
-
             if (is_scalar($attributeValue)) {
                 return trim($attributeValue);
-            } elseif (is_array($attributeValue)) {
+            }
+            if (is_array($attributeValue)) {
                 return implode(',', $attributeValue);
             }
         }
@@ -658,5 +675,57 @@ class Nosto_Tagging_Model_Meta_Product extends Nosto_Object_Product_Product
         }
 
         return $availability;
+    }
+
+    /**
+     * Adds selected attributes to all tags also in the custom fields section
+     *
+     * @param Mage_Catalog_Model_Product $product
+     * @param Mage_Core_Model_Store $store
+     * @return array
+     * @throws Nosto_NostoException
+     */
+    protected function buildCustomFields(
+        Mage_Catalog_Model_Product $product,
+        Mage_Core_Model_Store $store
+    )
+    {
+        $customFields = $this->loadCustomFields($product, $store);
+        $attributes = $this->getAttributesFromAllTags($store);
+        /* @var Mage_Catalog_Model_Resource_Eav_Attribute $productAttribute */
+        foreach ($product->getAttributes() as $key => $productAttribute) {
+            if (in_array($key, $attributes, false)) {
+                $attributeValue = $this->getAttributeValue($product, $key);
+                if ($attributeValue === null || $attributeValue === '') {
+                    continue;
+                }
+                $customFields[$key] = $attributeValue;
+            }
+        }
+        return $customFields;
+    }
+
+    /**
+     * Returns unique selected attributes from all tags
+     *
+     * @param Mage_Core_Model_Store $store
+     * @return array
+     * @throws Nosto_NostoException
+     */
+    protected function getAttributesFromAllTags(Mage_Core_Model_Store $store)
+    {
+        $attributes = array();
+        /* @var Nosto_Tagging_Helper_Data $nostoHelper */
+        $nostoHelper = Mage::helper("nosto_tagging");
+        foreach (Nosto_Tagging_Helper_Data::$validTags as $tagId) {
+            $attributesToTag = $nostoHelper->getAttributesToTag($tagId, $store->getId());
+            if (empty($attributesToTag) || !is_array($attributesToTag)) {
+                continue;
+            }
+            foreach ($attributesToTag as $tagAttribute) {
+                $attributes[] = $tagAttribute;
+            }
+        }
+        return array_unique($attributes);
     }
 }